[feat] finish Produto endpoints, need to fix reponse bugs

+++ src/main/java/ascii/ecommerce/demo/services/ProdutoService.java
@@ -1,7 +1,6 @@
 package ascii.ecommerce.demo.services;
 
 import ascii.ecommerce.demo.domain.produto.Produto;
-import ascii.ecommerce.demo.domain.produto.ProdutoPatchRequestDTO;
 import ascii.ecommerce.demo.domain.produto.ProdutoRequestDTO;
 import ascii.ecommerce.demo.domain.produto.ProdutoResponseDTO;
 import ascii.ecommerce.demo.repositories.ProdutoRepository;
@@ -30,20 +29,27 @@ public class ProdutoService {
     }
 
     public ProdutoResponseDTO getProdutoById(int id){
-        return this.produtoRepository
-                .findById(id)
-                .map(produto -> new ProdutoResponseDTO(produto))
-                .get();
+        if(this.produtoRepository.existsById(id)){
+            return this.produtoRepository
+                    .findById(id)
+                    .map(produto -> new ProdutoResponseDTO(produto))
+                    .get();
+        }
+        throw new RuntimeException("ID inválido ou produto inexistente");
     }
 
     public ProdutoResponseDTO getProdutoByNome(String nome){
-        return this.produtoRepository
+        ProdutoResponseDTO produtoBuscado = this.produtoRepository
                 .findAll()
                 .stream()
                 .filter(produto -> produto.getNome().equalsIgnoreCase(nome))
                 .findFirst()
                 .map(produto -> new ProdutoResponseDTO(produto))
                 .get();
+        if(this.produtoRepository.existsById(produtoBuscado.id())){
+            return produtoBuscado;
+        }
+        throw new RuntimeException("Nome inválido ou produto inexistente"); // Não funciona -> Verificar
     }
 
     public List<ProdutoResponseDTO> getProdutoListByCategoria(int categoria_id){
@@ -58,6 +64,7 @@ public class ProdutoService {
 
         // POST
 
+            // Colocar ativo=true por default
     public void addNewProduto(ProdutoRequestDTO produtoRequestDTO){
         this.produtoRepository.save(new Produto(produtoRequestDTO));
     }
@@ -66,50 +73,61 @@ public class ProdutoService {
         // PUT
 
     public void updateProdutoById(int id, ProdutoRequestDTO produtoRequestDTO){
-        if(this.produtoRepository.existsById(id)){
-            Produto produtoUpdate = new Produto(produtoRequestDTO);
-            produtoUpdate.setId(id);
-            if(produtoRequestDTO.nome() == null) produtoUpdate.setNome(this.produtoRepository.findById(id).get().getNome());
-            if(produtoRequestDTO.descricao() == null) produtoUpdate.setDescricao(this.produtoRepository.findById(id).get().getDescricao());
-            if(produtoRequestDTO.preco() == null) produtoUpdate.setPreco(this.produtoRepository.findById(id).get().getPreco());
-            if(produtoRequestDTO.categoria_id() == null) produtoUpdate.setCategoria_id(this.produtoRepository.findById(id).get().getCategoria_id());
-            if(produtoRequestDTO.ativo() == null) produtoUpdate.setAtivo(this.produtoRepository.findById(id).get().isAtivo());
-            this.produtoRepository.save(produtoUpdate);
+        try{
+            if(this.produtoRepository.findById(id).isPresent()){
+                Produto produtoUpdate = new Produto(produtoRequestDTO);
+                produtoUpdate.setId(id);
+
+                Produto produtoASerAtualizado = this.produtoRepository.findById(id).get();
+
+                if(produtoUpdate.getNome() == null) produtoUpdate.setNome(produtoASerAtualizado.getNome());
+                if(produtoUpdate.getDescricao() == null) produtoUpdate.setDescricao(produtoASerAtualizado.getDescricao());
+                if(produtoUpdate.getPreco() == null) produtoUpdate.setPreco(produtoASerAtualizado.getPreco());
+                if(produtoUpdate.getImagem() == null) produtoUpdate.setImagem(produtoASerAtualizado.getImagem());
+                if(produtoUpdate.getCategoria_id() == null) produtoUpdate.setCategoria_id(produtoASerAtualizado.getCategoria_id());
+                if(produtoUpdate.isAtivo() == null) produtoUpdate.setAtivo(produtoASerAtualizado.isAtivo());
+
+                this.produtoRepository.save(produtoUpdate);
+            }
+            else
+                throw new RuntimeException("ID inválido ou produto inexistente");
         }
-        // Retornar 400 se não existir
-        // Corrigir pois pode dar update só de um atributo, criando assim um produto com atributos vazios
+        catch (RuntimeException idInvalido){}
     }
+
 
     public void updateProdutoByNome(String nome, ProdutoRequestDTO produtoRequestDTO){
-        if(this.produtoRepository.findByNome(nome) != null){
-            Produto produtoUpdate = new Produto(produtoRequestDTO);
-            produtoUpdate.setId(this.produtoRepository.findAll().stream().filter(produto -> produto.getNome().equalsIgnoreCase(nome)).findFirst().get().getId());
-            if(produtoRequestDTO.nome() == null) produtoUpdate.setNome(this.produtoRepository.findByNome(nome).getNome());
-            if(produtoRequestDTO.descricao() == null) produtoUpdate.setDescricao(this.produtoRepository.findByNome(nome).getDescricao());
-            if(produtoRequestDTO.preco() == null) produtoUpdate.setPreco(this.produtoRepository.findByNome(nome).getPreco());
-            if(produtoRequestDTO.categoria_id() == null) produtoUpdate.setCategoria_id(this.produtoRepository.findByNome(nome).getCategoria_id());
-            if(produtoRequestDTO.ativo() == null) produtoUpdate.setAtivo(this.produtoRepository.findByNome(nome).isAtivo());
-            this.produtoRepository.save(produtoUpdate);
+        try{
+            if(this.produtoRepository.findByNome(nome) != null){
+                Produto produtoUpdate = new Produto(produtoRequestDTO);
+
+                Produto produtoASerAtualizado = this.produtoRepository.findAll().stream().filter(produto -> produto.getNome().equalsIgnoreCase(nome)).findFirst().get();
+
+                produtoUpdate.setId(produtoASerAtualizado.getId());
+
+                if(produtoUpdate.getNome() == null) produtoUpdate.setNome(produtoASerAtualizado.getNome());
+                if(produtoUpdate.getDescricao() == null) produtoUpdate.setDescricao(produtoASerAtualizado.getDescricao());
+                if(produtoUpdate.getPreco() == null) produtoUpdate.setPreco(produtoASerAtualizado.getPreco());
+                if(produtoUpdate.getImagem() == null) produtoUpdate.setImagem(produtoASerAtualizado.getImagem());
+                if(produtoUpdate.getCategoria_id() == null) produtoUpdate.setCategoria_id(produtoASerAtualizado.getCategoria_id());
+                if(produtoUpdate.isAtivo() == null) produtoUpdate.setAtivo(produtoASerAtualizado.isAtivo());
+
+                this.produtoRepository.save(produtoUpdate);
+            }
+            else {
+                throw new RuntimeException("ID inválido ou produto inexistente");
+            }
         }
-    }
-
-
-        // PATCH -> Update parcial
-    public void updatePartialProdutoById(int id, ProdutoPatchRequestDTO produtoPatchRequestDTO){
-        //Inserir Try-Catch
-        Produto produtoBuscado = this.produtoRepository.findById(id).orElseThrow(() -> new RuntimeException("Produto não cadastrado"));
-        if(produtoPatchRequestDTO.nome() != null) produtoBuscado.setNome(produtoPatchRequestDTO.nome());
-        if(produtoPatchRequestDTO.descricao() != null) produtoBuscado.setDescricao(produtoPatchRequestDTO.descricao());
-        if(produtoPatchRequestDTO.preco() != null) produtoBuscado.setPreco(produtoPatchRequestDTO.preco());
-        if(produtoPatchRequestDTO.imagem() != null) produtoBuscado.setImagem(produtoBuscado.getImagem());
-        if(produtoPatchRequestDTO.categoria_id() != null) produtoBuscado.setCategoria_id(produtoBuscado.getCategoria_id());
-        if(produtoPatchRequestDTO.ativo() != null) produtoBuscado.setAtivo(produtoBuscado.isAtivo());
-        
+        catch (RuntimeException idInvalido){}
     }
 
         // DELETE
 
     public void deleteProdutoById(int id){
-        this.produtoRepository.deleteById(id);
+        try{
+            if(this.produtoRepository.existsById(id)) this.produtoRepository.deleteById(id);
+            else throw new RuntimeException ("ID inválido ou produto inexistente");
+        }
+        catch (RuntimeException idInvalido){}
     }
 }
